+++ relayer/src/chain/cosmos.rs
@@ -14,13 +14,14 @@ use tendermint::block::Height;
 use tendermint::consensus::Params;
 use tendermint_light_client::types::LightBlock as TMLightBlock;
 use tendermint_proto::Protobuf;
+use tendermint_rpc::{Client, endpoint::broadcast::tx_commit::Response, HttpClient, Order};
 use tendermint_rpc::query::Query;
-use tendermint_rpc::{endpoint::broadcast::tx_commit::Response, Client, HttpClient, Order};
 use tokio::runtime::Runtime as TokioRuntime;
 use tonic::codegen::http::Uri;
 
 use ibc::downcast;
 use ibc::events::{from_tx_response_event, IBCEvent};
+use ibc::Height as ICSHeight;
 use ibc::ics02_client::client_def::{AnyClientState, AnyConsensusState};
 use ibc::ics03_connection::raw::ConnectionIds;
 use ibc::ics04_channel::channel::QueryPacketEventDataRequest;
@@ -30,20 +31,16 @@ use ibc::ics07_tendermint::consensus_state::ConsensusState as TMConsensusState;
 use ibc::ics07_tendermint::header::Header as TMHeader;
 use ibc::ics23_commitment::commitment::CommitmentPrefix;
 use ibc::ics23_commitment::merkle::convert_tm_to_ics_merkle_proof;
+use ibc::ics24_host::{IBC_QUERY_PATH, Path};
 use ibc::ics24_host::identifier::{ChainId, ChannelId, ClientId, ConnectionId};
 use ibc::ics24_host::Path::ClientConsensusState as ClientConsensusPath;
 use ibc::ics24_host::Path::ClientState as ClientStatePath;
-use ibc::ics24_host::{Path, IBC_QUERY_PATH};
-use ibc::Height as ICSHeight;
 // Support for GRPC
 use ibc_proto::cosmos::auth::v1beta1::{BaseAccount, QueryAccountRequest};
 use ibc_proto::cosmos::base::v1beta1::Coin;
-use ibc_proto::cosmos::tx::v1beta1::mode_info::{Single, Sum};
 use ibc_proto::cosmos::tx::v1beta1::{AuthInfo, Fee, ModeInfo, SignDoc, SignerInfo, TxBody, TxRaw};
-use ibc_proto::ibc::core::channel::v1::{
-    PacketState, QueryConnectionChannelsRequest, QueryPacketAcknowledgementsRequest,
-    QueryPacketCommitmentsRequest, QueryUnreceivedAcksRequest, QueryUnreceivedPacketsRequest,
-};
+use ibc_proto::cosmos::tx::v1beta1::mode_info::{Single, Sum};
+use ibc_proto::ibc::core::channel::v1::{PacketState, QueryChannelsRequest, QueryConnectionChannelsRequest, QueryPacketAcknowledgementsRequest, QueryPacketCommitmentsRequest, QueryUnreceivedAcksRequest, QueryUnreceivedPacketsRequest};
 use ibc_proto::ibc::core::commitment::v1::MerkleProof;
 use ibc_proto::ibc::core::connection::v1::QueryConnectionsRequest;
 
@@ -52,8 +49,8 @@ use crate::config::ChainConfig;
 use crate::error::{Error, Kind};
 use crate::event::monitor::{EventBatch, EventMonitor};
 use crate::keyring::store::{KeyEntry, KeyRing, KeyRingOperations, StoreBackend};
-use crate::light_client::tendermint::LightClient as TMLightClient;
 use crate::light_client::LightClient;
+use crate::light_client::tendermint::LightClient as TMLightClient;
 
 use super::Chain;
 
@@ -579,7 +576,7 @@ impl Chain for CosmosSDKChain {
         Ok((pc, height))
     }
 
-    /// Queries the packet commitment hashes associated with a channel.
+    /// Queries the unreceived packet sequences associated with a channel.
     fn query_unreceived_packets(
         &self,
         request: QueryUnreceivedPacketsRequest,
@@ -637,7 +634,7 @@ impl Chain for CosmosSDKChain {
         Ok((pc, height))
     }
 
-    /// Queries the packet commitment hashes associated with a channel.
+    /// Queries the unreceived acknowledgements sequences associated with a channel.
     fn query_unreceived_acknowledgements(
         &self,
         request: QueryUnreceivedAcksRequest,
@@ -751,6 +748,36 @@ impl Chain for CosmosSDKChain {
             .connections
             .iter()
             .filter_map(|ic| ConnectionId::from_str(ic.id.as_str()).ok())
+            .collect();
+
+        Ok(ids)
+    }
+
+    fn query_channels(&self, request: QueryChannelsRequest) -> Result<Vec<ChannelId>, Error> {
+        crate::time!("query_connections");
+
+        let grpc_addr =
+            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
+        let mut client = self
+            .block_on(
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+            )
+            .map_err(|e| Kind::Grpc.context(e))?;
+
+        let request = tonic::Request::new(request);
+
+        let response = self
+            .block_on(client.channels(request))
+            .map_err(|e| Kind::Grpc.context(e))?
+            .into_inner();
+
+        // TODO: add warnings for any identifiers that fail to parse (below).
+        //      similar to the parsing in `query_connection_channels`.
+
+        let ids = response
+            .channels
+            .iter()
+            .filter_map(|ch| ChannelId::from_str(ch.channel_id.as_str()).ok())
             .collect();
 
         Ok(ids)